move code from static folder to src using module system
