+++ src/.eslintrc.json
@@ -1,0 +1,15 @@
+{
+  "env": {
+    "browser": true,
+    "node": true
+  },
+  "extends": "eslint:recommended",
+  "parserOptions": {
+    "ecmaVersion": 2015,
+    "sourceType": "module"
+  },
+  "rules": {
+      "semi": "off",
+      "no-unused-vars": ["error", { "argsIgnorePattern": "^_" }]
+  }
+}
