+++ src/constants.js
@@ -1,0 +1,10 @@
+var constants = {
+  MMS_DOMAIN: 'https://mms.sp-prod.net',
+  CMP_ORIGIN: 'https://sourcepoint.mgr.consensu.org',
+  MSG_SCRIPT_URL: 'https://dialogue.sp-prod.net/messagingWithoutDetection.js'
+};
+
+export default constants;
+export var MMS_DOMAIN = constants.MMS_DOMAIN;
+export var CMP_ORIGIN = constants.CMP_ORIGIN;
+export var MSG_SCRIPT_URL = constants.MSG_SCRIPT_URL;
