+++ static/index.html
@@ -8,12 +8,7 @@
     </script>
   </head>
   <body style="overflow-y: hidden; height: 100vh;">
-    <script src="./on_document_ready.js" type="text/javascript"></script>
-    <script src="./constants.js" type="text/javascript"></script>
-    <script src="./amp_client.js" type="text/javascript"></script>
-    <script src="./sourcepoint_client.js" type="text/javascript"></script>
-    <script src="./sourcepoint_config.js" type="text/javascript"></script>
-    <script src="./load_message.js" type="text/javascript"></script>
+    <script src="./dist/sourcepoint_amp.js"></script>
   </body>
   <div id="__genieContainer" style="all: initial;"></div>
 </html>
